% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioassays.R
\docType{package}
\name{bioassays}
\alias{bioassays}
\title{Bioassays: A Package for Analyzing Multi Well Plate Bioassay's.}
\description{
The bioassays package provides three categories of important functions for
extracting, formatting, plotting and analyzing.
}
\section{Extract functions. The function that help to extract information from file name is \code{\link{extract_filename}}}{
NA
}

\section{Format functions. The function to format data are \code{\link{data2plateformat}},\code{\link{plate2df}},\code{\link{matrix96}},\code{\link{plate_metadata}}}{
NA
}

\section{Plot functions. The function to plot is\code{\link{heatplate}}}{
NA
}

\section{Analysis functions. The function to analyse data are\code{\link{reduceblank}},\code{\link{estimate}},\code{\link{dfsummary}},\code{\link{pvalue}}}{
NA
}

