% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafile96.R
\docType{data}
\name{metafile96}
\alias{metafile96}
\title{Metadata of 96 Well Plate.}
\format{
{ A data frame with 96 rows and 7 variables.
 \describe{
 \item{row}{Row number of multi well well plate.}
 \item{col}{Column number of multi well plate.}
 \item{position}{Well position address of multi well plate.}
 \item{id}{Type of sample. 'STD' represent standards, 'sample' represent samples.}
 \item{type}{Type of sample. 'STD1','STD2','STD3' etc represent different standards, 'S1','S2','S3' represent different samples.}
 \item{concentration}{concentration of different standards (mg/ml).}
 \item{dilution}{dilution of samples used for assay.}
 }
 }
}
\source{
User generated metadata of 96 well plate.
}
\usage{
metafile96
}
\description{
A dataset containing metadata.
}
\keyword{datasets}
