% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata384.R
\docType{data}
\name{rawdata384}
\alias{rawdata384}
\title{Rawdata of 384 well plate.}
\format{
A data frame with 16 rows and 25 columns: First column shows 'row name'.
}
\source{
Spectrophotometer output reading (OD) of 384 well plate.
}
\usage{
rawdata384
}
\description{
A dataset of spectrophotometer readings (OD).
}
\keyword{datasets}
