% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata96.R
\docType{data}
\name{rawdata96}
\alias{rawdata96}
\title{Rawdata of 96 well plate.}
\format{
A data frame with 8 rows and 13 columns: First column shows 'row name'.
}
\source{
Spectrophotometer output reading (OD) of 96 well plate.
}
\usage{
rawdata96
}
\description{
A dataset of spectrophotometer readings (OD).
}
\keyword{datasets}
