% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioassays.R
\docType{package}
\name{bioassays}
\alias{bioassays}
\title{Bioassays: A Package for Analyzing Multi Well Plate Bioassay's.}
\description{
The bioassays package provides three categories of important functions for extracting, formatting, plotting and analyzing.
}
\section{Extract functions}{

The function that help to extract information from file name is \cr
\code{\link{extract_filename}}'
}

\section{Format functions}{

The function to format data are \cr
\code{\link{data2plateformat}}\cr
\code{\link{plate2df}}\cr
\code{\link{matrix96}}\cr
\code{\link{plate_metadata}}
}

\section{Plot functions}{

The function to plot is\cr
\code{\link{heatplate}}
}

\section{Analysis functions}{

The function to analyse data are\cr
\code{\link{reduceblank}}\cr
\code{\link{estimate}}\cr
\code{\link{dfsummary}}\cr
\code{\link{pvalue}}
}

